renamed folder to AboutUs from aboutUs
